+++ src/NzbDrone.Core/Parser/Augmenters/AugmentWithMediaInfo.cs
@@ -27,24 +27,24 @@ namespace NzbDrone.Core.Parser.Augmenters
                 if (!(quality.Quality.Modifier == Modifier.BRDISK || quality.Quality.Modifier == Modifier.REMUX) &&
                     (quality.Quality.Source == Source.BLURAY || quality.Quality.Source == Source.TV ||
                      quality.Quality.Source == Source.WEBDL) &&
-                    !(quality.Quality.Resolution == (int)Resolution.R480p || quality.Quality.Resolution == (int)Resolution.R576p))
+                    !(quality.Quality.Resolution == Resolution.R480p || quality.Quality.Resolution == Resolution.R576p))
                 {
                     var width = mediaInfo.Width;
                     var existing = quality.Quality.Resolution;
 
                     if (width > 854)
                     {
-                        quality.Quality.Resolution = (int)Resolution.R720p;
+                        quality.Quality.Resolution = Resolution.R720p;
                     }
 
                     if (width > 1280)
                     {
-                        quality.Quality.Resolution = (int)Resolution.R1080p;
+                        quality.Quality.Resolution = Resolution.R1080p;
                     }
 
                     if (width > 1920)
                     {
-                        quality.Quality.Resolution = (int)Resolution.R2160p;
+                        quality.Quality.Resolution = Resolution.R2160p;
                     }
 
                     if (existing != quality.Quality.Resolution)
